+++ src/core/hw/tegra_x1/gpu/renderer/renderer_base.hpp
@@ -65,7 +65,7 @@ class RendererBase {
 
     // Buffer
     virtual BufferBase* CreateBuffer(const BufferDescriptor& descriptor) = 0;
-    virtual BufferBase* AllocateTemporaryBuffer(const usize size) = 0;
+    virtual BufferBase* AllocateTemporaryBuffer(const u32 size) = 0;
     virtual void FreeTemporaryBuffer(BufferBase* buffer) = 0;
 
     // Texture
